+++ test/nuclearComponent.spec.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react/addons'
+import React, { Component, createClass } from 'react/addons'
 import jsdomReact from './jsdomReact'
 import provideReactor from '../provideReactor'
 import nuclearComponent from '../nuclearComponent'
@@ -45,9 +45,8 @@ describe('nuclear component', () => {
 
     it('should provide bindings as props on mount', () => {
       @provideReactor
-      @nuclearComponent({
-        foo: 'bar',
-        test: 'lol',
+      @nuclearComponent(() => {
+        return { foo: 'bar', test: 'lol'}
       })
       class Parent extends Component {
         render() {
@@ -156,6 +155,60 @@ describe('nuclear component', () => {
     }
 
     React.render(<NoBindings reactor={fakeReactor}/>, div)
+    React.unmountComponentAtNode(div)
+  })
+
+  it('should not throw when using es5', () => {
+    const div = document.createElement('div')
+
+    const fakeReactor = {
+      evaluate() {},
+      observe() {
+        return () => {}
+      },
+    }
+
+    let SomeES5Stuff = createClass({
+      render() {
+        return <div/>
+      },
+    })
+
+    SomeES5Stuff = nuclearComponent(SomeES5Stuff, function() {
+      return { pass: 'through' }
+    })
+    SomeES5Stuff = provideReactor(SomeES5Stuff)
+
+    React.render(<SomeES5Stuff reactor={fakeReactor}/>, div)
+    React.unmountComponentAtNode(div)
+  })
+
+  it('should provide access to the props easily', () => {
+    const div = document.createElement('div')
+
+    const fakeReactor = {
+      evaluate(key) {
+        if (key !== 'through1337') {
+          throw new Error('wrong key')
+        }
+      },
+      observe() {
+        return () => {}
+      },
+    }
+
+    let SomeES5Stuff = createClass({
+      render() {
+        return <div/>
+      },
+    })
+
+    SomeES5Stuff = nuclearComponent(SomeES5Stuff, function(props) {
+      return { pass: 'through' + props.foo }
+    })
+    SomeES5Stuff = provideReactor(SomeES5Stuff)
+
+    React.render(<SomeES5Stuff foo="1337" reactor={fakeReactor}/>, div)
     React.unmountComponentAtNode(div)
   })
 })