+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0 (September 14, 2015)
+
+ - BREAKING CHANGE: nuclearComponent API matches nuclearMixin API for getDataBindings
+
+ now the nuclearComponent will always take a function that receives props (to avoid `this` bindings ambiguity and to match the constructor signature if using es6)
+
 ## 0.1.1 (August 10, 2015)
 
  - Fix documentation